refactor(users): 컨트롤러 DI 적용 및 클래스 기반 변경
