+++ src/v1/apis/users/users.controller.ts
@@ -1,15 +1,17 @@
 import { FastifyReply, FastifyRequest } from 'fastify';
 
-import { getUserParamsSchema } from '../users/user.schema.js';
+import { getUserParamsSchema } from './user.schema.js';
 import { ForbiddenException } from '../../common/exceptions/core.error.js';
 
-export async function findUserController(request: FastifyRequest, reply: FastifyReply) {
-  const params = getUserParamsSchema.parse(request.params);
-  const user = request.me;
+export default class UsersController {
+  async findUser(request: FastifyRequest, reply: FastifyReply) {
+    const params = getUserParamsSchema.parse(request.params);
+    const user = request.me;
 
-  if (user.id !== params.id) {
-    throw new ForbiddenException('You are not authorized to view this user');
+    if (user.id !== params.id) {
+      throw new ForbiddenException('You are not authorized to view this user');
+    }
+
+    reply.code(200).send('hello');
   }
-
-  reply.code(200).send('hello');
 }
